+++ src/nppi/nppi_arithmetic_operations/nppi_subc.cpp
@@ -281,7 +281,7 @@ NppStatus nppiSubC_32f_C1R_Ctx(const Npp32f * pSrc1, int nSrc1Step, const Npp32f
         return NPP_SIZE_ERROR;
     }
     
-    if (nSrc1Step < oSizeROI.width * sizeof(Npp32f) || nDstStep < oSizeROI.width * sizeof(Npp32f)) {
+    if (nSrc1Step < static_cast<int>(oSizeROI.width * sizeof(Npp32f)) || nDstStep < static_cast<int>(oSizeROI.width * sizeof(Npp32f))) {
         return NPP_STRIDE_ERROR;
     }
     
@@ -409,8 +409,8 @@ NppStatus nppiSubC_16u_C3RSfs_Ctx(const Npp16u * pSrc1, int nSrc1Step, const Npp
         return NPP_SIZE_ERROR;
     }
     
-    if (nSrc1Step < oSizeROI.width * 3 * sizeof(Npp16u) || 
-        nDstStep < oSizeROI.width * 3 * sizeof(Npp16u)) {
+    if (nSrc1Step < static_cast<int>(oSizeROI.width * 3 * sizeof(Npp16u)) || 
+        nDstStep < static_cast<int>(oSizeROI.width * 3 * sizeof(Npp16u))) {
         return NPP_STRIDE_ERROR;
     }
     
@@ -446,8 +446,8 @@ NppStatus nppiSubC_32f_C3R_Ctx(const Npp32f * pSrc1, int nSrc1Step, const Npp32f
         return NPP_SIZE_ERROR;
     }
     
-    if (nSrc1Step < oSizeROI.width * 3 * sizeof(Npp32f) || 
-        nDstStep < oSizeROI.width * 3 * sizeof(Npp32f)) {
+    if (nSrc1Step < static_cast<int>(oSizeROI.width * 3 * sizeof(Npp32f)) || 
+        nDstStep < static_cast<int>(oSizeROI.width * 3 * sizeof(Npp32f))) {
         return NPP_STRIDE_ERROR;
     }
